pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0", "polars>=1.3.0"]
build-backend = "maturin"

[project]
name = "dmpworks"
description = "DMP Tool Related Works Python Package"
requires-python = ">=3.12"
license = { text = "MIT license" }
dynamic = ["version"]
dependencies = [
    "polars>=1,<2",
    "pendulum>=3,<4",
    "tqdm>=4,<5",
    "opensearch-py>=2.8.0,<3",
    "pyarrow>=19,<21",
    "sqlmesh>=0.188.1,<1",
    "cyclopts>=3,<4",
    "pooch>=1,<2",
    "boto3>=1,<2",
    "orjson>=3,<4",
    "fold_to_ascii>=1,<=2",
    "rapidfuzz>=3,<4",
    "beautifulsoup4>=4,<5",
    "pydantic>=2,<3",
    "jsonlines>=4,<5",
    "duckdb>=1.4.0,<2",
    "pymysql>=1.1.2,<2"
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9", # testing
    "pytest-mock>=3,<4", # testing
    "black>=25,<26", # code formatting
    "maturin>=1.0,<2.0", # Build system
    "ruff>=0.11,<0.12", # TODO: evaluate if needed
    "mypy>=1,<2", # TODO: evaluate if needed
    "vcrpy>=7,<8"
]

[project.scripts]
dmpworks = "dmpworks.cli:main"

[tool.maturin]
module-name = "dmpworks.polars_expr_plugin._internal"
python-source = "python"
include = ["python/**/*"]

[[tool.mypy.overrides]]
module = "polars.utils.udfs"
ignore_missing_imports = true

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
